requirements.txt
streamlit
langchain
langchain-community
langchain-core
python-dotenv
